Add SSL Options
Added the ability to define SSL frontend pools. The first attempt as of
now is to add the ability to generate a self signed SSL cert for demo
purposes. The ability to also sync this certificate to other load
balancers so they are all using the same SSL certs. Also cleaned up
vars, updated example playbook, and etc.

This commit resolves #5

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>